@@ -28,7 +28,6 @@ classifiers =
     Operating System :: OS Independent
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: Implementation :: CPython
@@ -41,7 +40,7 @@ package_dir=
 	=src
 packages = find:
 zip_safe = False
-python_requires = >=3.5
+python_requires = >=3.6,<3.8
 install_requires =
     astropy>=3.1,<4.*
     astroquery>=0.3.9
@@ -62,9 +61,9 @@ console_scripts = poliastro = poliastro.cli:main
 
 [options.extras_require]
 jupyter = notebook; ipywidgets>=7.0
-cesium = czml3~=0.1.2 ; python_version>='3.6'
+cesium = czml3~=0.1.2
 dev =
-    black ; python_version>='3.6'
+    black
     coverage
     ipykernel
     ipython>=5.0
